+++ .github/workflows/code-quality.yaml
@@ -83,7 +83,7 @@ jobs:
 
     - name: install python dependencies
       id: pip
-      run: python -m pip install -e .[dev]
+      run: python -m pip install -e .[dev,docs]
 
     - name: pytest
       id: pytest
@@ -92,12 +92,12 @@ jobs:
     - name: coverage (not windows)
       if: runner.os != 'Windows'
       id: coverage-not-windows
-      run: python -m coverage report --omit=**/*messagequeue*,**/mq/* --fail-under=95
+      run: python -m coverage report --omit=**/*messagequeue*,**/mq/* --fail-under=90
 
     - name: coverage (windows)
       if: runner.os == 'Windows'
       id: coverage-windows
-      run: python -m coverage report --omit=**/*messagequeue*,**/mq/* --fail-under=90
+      run: python -m coverage report --omit=**/*messagequeue*,**/mq/* --fail-under=89
 
   test-e2e:
     name: 'e2e tests ${{ matrix.run_mode }}'
