+++ grizzly/steps/scenario/tasks/conditional.py
@@ -41,7 +41,7 @@ def step_task_conditional_if(context: Context, condition: str, name: str) -> Non
 @then(u'if condition is false, execute these tasks')
 def step_task_conditional_else(context: Context) -> None:
     """
-    Changes the state of {@pylink grizzly.tasks.conditional} task instance created by {@pylink grizzly.steps.scenario.tasks.step_task_conditional_if}
+    Changes the state of {@pylink grizzly.tasks.conditional} task instance created by {@pylink grizzly.steps.scenario.tasks.conditional.step_task_conditional_if}
     so that any following tasks will be run when `condition` is false.
 
     See {@pylink grizzly.tasks.conditional} task documentation for more information.
@@ -67,7 +67,7 @@ def step_task_conditional_else(context: Context) -> None:
 @then(u'end condition')
 def step_task_conditional_end(context: Context) -> None:
     """
-    Closes the {@pylink grizzly.tasks.conditional} task instance created by {@pylink grizzly.steps.scenario.tasks.step_task_conditional_if}.
+    Closes the {@pylink grizzly.tasks.conditional} task instance created by {@pylink grizzly.steps.scenario.tasks.conditional.step_task_conditional_if}.
     This means that any following tasks specified will not be part of the conditional.
 
     See {@pylink grizzly.tasks.conditional} task documentation for more information.
