+++ grizzly/tasks/wait_explicit.py
@@ -1,10 +1,10 @@
 """
-@anchor pydoc:grizzly.tasks.explicit_wait Explicit Wait
+@anchor pydoc:grizzly.tasks.wait_explicit Explicit Wait
 This task executes a `gevent.sleep` and is used to manually create delays between steps in a scenario.
 
 ## Step implementations
 
-* {@pylink grizzly.steps.scenario.tasks.wait_explicit.step_task_wait_seconds}
+* {@pylink grizzly.steps.scenario.tasks.wait_explicit.step_task_wait_explicit}
 
 ## Arguments
